Add `traffic.rs`:
`TrafficState`: has four vectors for cars in arriving lanes, and four vectors for cars in leaving lanes.

It has an array of four booleans which indicates if a new car can be placed in the array. Can use this to implement a delay, or combine it for use with the capacity of the vectors.

It has an array of the four traffic lights. `true` is a green light.

`add_car`: adds a car if the vector is not full. It has basic implementation of random direction.