Add `car.rs`:
`ComingFrom` is an enum which signifies the direction where the car is coming from. Right now `Random` is implemented in it, but should later be moved when `Controls` are implemented.

`Going` is an enum which symbolizes if the car will make a turn, or keep going straight.

`Car` is a struct which contains `ComingFrom`, `Going`, x position, y position, and a bool which tells it if it is moving.

It has basic implementation of moving the cars with the `move_car` method.